Add transactional support to Kafka publisher
This commit introduces transactional support to the KafkaMessagePublisher class. The new class, KafkaTransactionalMessagePublisher, extends the existing KafkaMessagePublisher and overrides the send methods to ensure that messages are sent within a transaction. This is important to ensure that either all messages are successfully sent, or none are in case of an error.

The new class also maintains a map of producers for each thread, ensuring that each thread has its own producer. This is necessary because Kafka's transactional producers are not thread-safe.

Resolves: TTMUUD-2303
Related: TTMUUD-2282